Add files via upload
The role of a photo management application is to organize photographs so that they can
be easily accessed. In order to help organize photos, the user can provide tags to describe
the content of the photos. A tag is a keyword associated with a photo,
In order to accelerate the search for photos, it is possible to store the tags in an inverted
index. The idea is that instead of having the photos point to the tags, the inverted index
will store all the tags, and each tag will point to all the photos that contain it